Eliminate inconsistency in process min heap
The previous commit allowed the process min heap to lose
the min heap property for short periods. While this did not
cause problems in test cases, it might lead to problems in
the future. Here this problem is resolved by calculating
an updated process cost when thread 0 is assigned a new chunk,
but not changing the cost associated with the process in
the min heap until after new chunks are also assigned to the
other threads.

[BFB]