SCREAM: moved surface coupling files to control folder and worked on their impl.
They should be in that folder, cause the AD will need those classes, and
it is awkward to make the control folder depend on the mct_interface one.
Besides, surface_coupling is generic, and does not have any mct-specific
code in it.

Additionally, adapt surface coupling to comply with changes in Field class.