Rework snowcapping (from CESMv2)
If use_extrasnowlayers, then a new snowcapping routine will be
used, which is contained in SnowHydrologyMod.  The old (current)
snowcapping scheme touches multiple files, and causes a column
in a snow capped state to have a virtual roof over the snow.
Excess snow is sent either to a runoff term or an ice sheet source
term (positive surface mass balance), but new snow does not
accumulate at the surface.  This causes the snow to age as
if no new snowfall is occuring and leads to biases in its density.

The new routine, adapted from CLMv5, removes excess snow from
the bottom of the snow column, allowing the snowpack to refresh
from the top like it would in nature.  This commit places control
logic surrounding the relevant snow capping infrastructure so that
if use_extrasnowlayers is true, then the new routine will be toggled
while maintaining conservation of mass.