SCREAM: a few mods to field structures stack.
Most notably, added getter to FieldAllocProp, to retrieve the extent of
the allocation along the last dimension, in terms of the stored scalar type.
This might be different from the last dim in the field layout, in case of
padding.

Also, changed `tag2str` to `e2str`, which is what other enum->string functions
are called in scream.