Checkout E3SM upstream RRTMGP interface
Checkout E3SM versions of RRTMGP interface files, and make appropriate
changes to namelist variables and use case files to accommodate. There
is really no reason for us to use a custom version of the RRTMGP
interface, so we should move future development upstream to the E3SM
repo and adopt the changes there with upstream merges.