add -Mnovect when compiling cice using PGI on compy
An F case

 -res ne4pg2_ne4pg2 -compset FC5AV1C-L -mach compy -compiler pgi

is not BFB with respect to changes in PE layout, for example,
64x1 compared to 96x1, with the current optimization flags.
Reproducibility can be restored by compiling the CICE component
with the flag -Mnovect. This is implemented by adding the rule

    <append COMP_NAME="cice">-Mnovect</append>

to config_compilers.xml, which adds

  if("${COMP_NAME}" STREQUAL "cam")
    set(FFLAGS "${FFLAGS} -Mnovect")
  endif()

to Macros.cmake and

  ifeq ($(COMP_NAME),cice)
    FFLAGS := $(FFLAGS) -Mnovect
  endif

Macros.make in a case directory.

[Non-BFB]