Have only process 0 write out chunk cost data
The current algorithm used in print_cost_p has each process, in turn,
open the file atm_chunk_cost.txt, write out the measured cost
assocated with each chunk to the end of the file, and then close the
file. This approach is efficient with respect to memory usage, but is
costly in terms of runtime for large process counts on systems such as
Cori-KNL at NERSC.

Here an alternative algorithm is implemented. All data
to be written out is first gathered to the ATM masterproc. The
file atm_chunk_cost.txt is then opened/created once, the data
is written out by masterproc, and the file is then closed. The memory
requirements are higher with this approach, but, at scale, this
alternative approach has been observed to be almost 300 times faster
on Cori-KNL .

[BFB]