Hommexx: fix to HybridVCoord, to avoid read OOB "error"
The OOB read is "innocuous", since the garbage values only affect the
padding in the last pack of the hyai_delta/hybi_delta views in hvcoord.
Still, it is not right, and if someone is using valgrind, these errors
might cause a waste of debug time.

The solution is simple: when using 2 views for the same array (one packed,
and one unpacked), we should allocate the longest (the packed one), and
make the other (the unpacked one) "view" it. This way, we are sure we will
always read inside bounds of memory we allocated.

Also, added a small unit test, to verify that compute_deltas in the hvcoord
class does the right thing.