Create new method for distributing calving
This routine takes a calving velocity and converts it to a
thickness removal.  The basic idea is to multiply the calving velocity
by the edge length and height to get a volume flux.  However, the details
are complicated by the unstructured Voronoi mesh and the use of partially-filled
non-dynamic floating grid cells.
The algorithm calculates the required calving volume flux based on edges, and then
applies the volume removal to grid cells, converting that to a thickness removal using areaCell.
This process occurs in phases.

The first phase handles non-dynamic floating cells. This is done by calculating the required calving flux
at the edges between those non-dynamic floating cells and the open ocean.
The calving velocity and ice thickness are copied from the upstream dynamic
floating cells, because these quantities do not have valid values on the non-dynamic cells.
After calculating the calving flux determined at each of these edges, the non-dyamic floating cells
are looped over, and volume is marked for removal based on the calving flux on the edges of each cell.

The final phase handles dynamic floating cells that are eligible for calving.
There are two ways this can happen - either they are adjacent to the open ocean (no buffer of
non-dynamic cells present), or they are adjacent to non-dynamic floating cells that had their
entire volume removed by calving, but there still is calving demand remaining.
The first case is handled easily by calculating calving flux on edges between dynamic floating cells
and open ocean using the calving velocity and ice thickness of the dynamic cell itself.
In the second case, "leftover" required calving flux is distributed from a non-dynamic cell that
has been completely "drained" to any edges between that non-dynamic cell and dynamic cells upstream.
This is done by weighting the flux to be distrbuted by the length of each such edge in the direction
perpendicular to the calving flux.