Assume calving is normal to velocity
This improvement assumes the calving velocity is normal to the advective
velocity by scaling edge length by the dot product between the flow
velocity direction and the edge normal direction.

Note that this implementation does not check for positive or negative
which would require a more careful calculation.  However the situation
of calving operating in the reverse direction as the flow direction
seems unlikely.  This perhaps should be investigated more carefully.