Make BG and IG compsets used Static MALI (but with Albany build)
This is to have these compsets by default have ice sheet evolution
disabled (because it is still problematic), but allow it to be enabled
with user_nl changes so that runs with active ice sheets can be
performed easily.  In the future when ice evolution is happening
reliably, this commit can be reversed.