Bug fixes to calving scheme for multiple procs
This commit makes a number of bug fixes to get correct results for multiple processors
and some other adjustments:

* adds calvedVolume variables to Registry
* Adds halo update to uncalvedVolumeNonDynCell after phase 1
* In phase 2, loop over nCells instead of nCellsSolve so you can copy unused calving from the halo region
into the owned cells
* In phase 2, perform halo update on requiredCalvingVolumeDynEdge.  (Not sure if this one is necessary)
* Adjust logic in applying calving in phase 2 to loop over all dynamic ice cells that are below sea level.
There is no harm on looping over extra cells, because calving will only occur if the edges there have a need
for calving.  There is some cost to evaluating the rest of this section at additional cells.  (Not sure if
this change is necessary.)