Add input class to scorpio I/O management.
This commit adds a AtmosphereInput class which is a trimmed down version
of the AtmosphereOutput class.

Input does not require as many steps as output, nor does it typically
require management of the file over more than one timestep.  As a result,
this class is much simpler and does not need to be used by the OutputManager.

I note here that a future commit should create an I/O class with
AtmosphereOutput and AtmosphereInput as sub-classes.

If time-managed input is required at a later developmental stage than it should
be possible to include it into the OutputManager.

Finally, this commit introduces a test for input in the scorpio_yaml test.  Currently,
this test is rather elementary.  A subsequent commit will improve the test.