Defense against use of user_nl_cam after renaming cam to eam
When recycling a run script for a new case, one may not realize
the case specific namelist settings may still go to user_nl_cam.
This change is to force users to rename it to user_nl_eam before
proceeding, if user_nl_cam is still in use, or present in the
case directory. Similar is done for user_nl_clm.

[BFB]