Set fill values for radiation outputs
Set the flag_xyfill flag for radiation fields added to history files in
addfld calls. This fixes the problem where radiation fields can contain
zeros if the output frequency is higher (more often) than the radiative
update frequency. This is BFB, except for the case where 'HR' is added
to output fields because we were mistakenly NOT computing and outputting
this field for RRTMGP.