Homme: adding HOMME_WITHOUT_PIOLIBRARY to compile definitions for executables
This is not needed for homme standalone builds. But if homme is built inside
another project, we cannot know for sure that the place where the macros
to create the execs are called from. If they are called from _outside_
the homme subfolder, all the macros that were added with ADD_DEFINITION(..)
(like this one) are no longer visible, and need to be manually added again.