Homme: adding HOMME_WITHOUT_PIOLIBRARY to compile definitions of executables
This is not needed for homme standalone builds. But if homme is built inside
another project, we cannot know in advance the place where the macros
to create the execs are called from. If they are called from *outside*
the homme subfolder, all the compiler macros that were added with ADD_DEFINITION(..)
(like this one) are no longer visible, and need to be manually added again.
This is because ADD_DEFINITION (like INCLUDE_DIRECTORIES and others0 only
affects targets built in the current folder and its subdirectories.