Homme: store blas/lapack libs location in an internal cache var
As explained in the inline comment, this is needed if homme is used as a
subproject in another cmake project. If so, when calling macros to create
execs/libs from *outside* the homme subfolders, the cmake vars storing
the location of libs found during homme's configuration (via calls to
find_package(...)) would not otherwise be available anymore, because of
cmake's variables scope. Stuffing those vars in the cache fixes this.