Enable the geotile manager to supersample tiles
This commit enables the supersampling of tiles by updating the geotile manager
tile_to_latlon function to subdivide the pixels of a tile if a supersampling
factor is given in the function call. The code that calls tile_to_latlon will
need to pass in supersampled coordinates if supersample_fac > 1.

This is currently needed for the interpolation of maximum snow albedo, but will
also be needed for future supersampling when other datasets become coarser than
grid spacing.