Make changes needed to get tests working on LC machines.
This commit adjusts the cmake flags to extend the generality in how
mpirun is used in testing.  The immediate impact is that the two LC
machine syrah and quartz will be able to run tests using srun, since
mpirun is not available.

There are also adjustments to fix lassen machine configurations.