Remove scalefactor from read_geogrid calls in init_atm_gwd
This commit removes the scalefactor argument from the calls to read_geogrid in
mpas_init_atm_gwd. Although currently the scalefactor is 1.0 and has no effect
on the tiles after they are read in, the tile is scaled regardless as a
placeholder for future datasets which may need to be scaled by a different
value.