SCREAM: fixes to atm inputs initialization logic.
If the FieldInitializer inits a field on the ref grid, all works as before.
However, if it inits a field F not on the ref grid, the driver will request
that it also inits F_ref, a copy on the ref grid. To make this easy, the AD will
provide to the initializer a remapper from current to ref grid, so that
if the initializer is not able to init the ref field by itself, it can
do so by simply remapping F to F_ref, by calling the remapper.