Replace array reference argument with a scalar
The call to

  call proc_heap_adjust(inv_proc_heap(ntmp2), proc_heap_remove,   &
                        max_nvsmptasks, proc_heap_len, proc_heap, &
                        proc_cost, inv_proc_heap)

in phys_grid.F90 where

  subroutine proc_heap_adjust(updated, remove, heap_dlen, &
                              heap_len, heap, cost, inv_heap)
  ...
  integer, intent(in)     :: updated
  ...
  integer, intent(inout)  :: inv_heap(0:npes-1)

modifies 'inv_heap' in the routine proc_heap_adjust, and changes the
value at the location 'inv_proc_heap(ntmp2)'. Compilation with
optimization greater than -O0 apparently caches this value (since it
is declared 'intent(in)'), so 'updated' is not changed within the
scope of the 'proc_heap_adjust' routine. -O0 compilation does not,
changing the value on 'updated' internally when 'inv_proc_heap(ntmp2)'
is changed, which causes a failure. This bug is fixed by changing
the call as follows:

  integer :: updated
  ...
  updated = inv_proc_heap(ntmp2)
  call proc_heap_adjust(updated, proc_heap_remove,   &
                        max_nvsmptasks, proc_heap_len, proc_heap, &
                        proc_cost, inv_proc_heap)

[BFB]