SCREAM: bug fix in FieldAllocProp
The layout should be passed at commit time, not construction. In fact, when
the alloc properties are constructed, it might be that the field layout
was not yet fully init-ed. Since we don't need the layout in any other place
other than the commit method, we might as well simply pass it at commit time.