In interface, only use temperatures where MPAS calculates them
This commit adjusts the Interface to only use MPAS temperatures values
from grid cells where the thickness exceeds the thickness limit used by
the thermal solver.  This avoids inconsistencies between what are
considered valid temperatures that can lead to bad temperature values.

Note that this condition is followed even if the MPAS thermal solver is
off.  So config_thermal_thickness has meaning to the interface even if
the thermal solver is not being used.