Add support for chrysalis to provenance.py
Created a new CIME branch (worleyph/chrysalis_support)
with a modified provenance.py that adds the chrysalis
system to the list of supported systems.

The CIME branch also (a) removes support for mira and
titan systems, and (b) fixes archiving of batch job
output files for slurm systems anvil, compy, cori-haswell,
and cori-knl.

[BFB]