Fixes a NBFB issue when clubb_use_sgv and zmconv_tp_fac are used
The model produces NBFB answers when run with phys_loadbalance =0
and phys_loadbalance=2 with the following settings in atm_in:

clubb_use_sgv          = .true.
zmconv_tp_fac          = 2.0D0

The reason is that the "tpert" variable is used in zm_convr subroutine
directly without collecting the "gathered" columns of "tpert". To
fix this, a new variable "tpertg" is declared which has the "gathered"
columns of "tpert". "tpertg" is being used instead of "tpert" for
cloud properties calculations.