SCREAM: minor mods to FieldPropertyCheck
The scalar type of the field during check and repair can be made
more specific. In particular, during check, it can be const, while
during repair it *must* be non-const.

Also, property_check_type in the Field class is an alias to
FieldPropertyCheck<non_const_RT>, to facilitaty copy of Field
object (non_const_RT is an alias to RealType without any const
qualifier).