SCREAM: better structuring requests for field groups
Added a mini sturct to hold the request info, which now also
include a pack size, so that an atm proc can request a group
of fields, and be able to use them with a pack size>1.

Added registration method to FieldRepository to allow specification
of value type specs (the pack size) as a runtime arg. This allows
the AD to treat the requests at runtime without compile time
knowledge of the requests.