Major update to include HB scheme in vdiff
All tests with the new MMF surface module that dumped the surface fluxes
into the atmosphere at various places in the loop without diffusion
failed. This is pretty dissappointing because it seems were are
completely reliant on some form of vertical diffusion outside the CRM.
In order to embrace this reality Icopied the HB scheme code in the crm
folder and restructured things so that this would always be used when
running the MMF. A lot of experimental code and ifdefs are still in
place currently. More testing is needed to determine if we should run
with the turbulent mountain stress on as well because this seems to be a
good way to stabilize the model at ne30pg2.

	modified:   components/eam/src/physics/crm/cloud_diagnostics.F90
	modified:   components/eam/src/physics/crm/crm_history.F90
	modified:   components/eam/src/physics/crm/crm_physics.F90
	modified:   components/eam/src/physics/crm/diffusion_solver.F90
	deleted:    components/eam/src/physics/crm/eddy_diff.F90
	renamed:    components/eam/src/physics/crm/mmf_hb_diff.F90 -> components/eam/src/physics/crm/hb_diff.F90
	modified:   components/eam/src/physics/crm/physpkg.F90
	modified:   components/eam/src/physics/crm/rrtmgp/radiation.F90
	renamed:    components/eam/src/physics/crm/mmf_vertical_diffusion.F90 -> components/eam/src/physics/crm/vertical_diffusion.F90