Extend ablationVelocity from dynamic to non-dynamic cells
This addresses a problem that stemmed from bad strain rate values in non-dynamic
cells being used to calculate calvingVelocity. This change sets the ablationVelocity
in non-dynamic cells to be the average of neigboring dynamic cells. In the case of a
stranded non-dynamic cell (i.e., one with only non-dynamic neighbors), thicknessForAblation
and ablationVelocity are both calculated from the neighboring non-dynamic cells