Add option for implicit surface flux calculations
This update adds an option for the surface models to calculate surface
stresses output to the atmosphere using an implicit method. This option
is controlled by the new ATM_FLUX_INTEGRATION_METHOD variable in
env_run.xml. If this option is set to "explicit", the original flux
calculations are used. If this option is set to "implicit_stress", the
new implicit stress code is used.

The explicit (default) setting is bit-for-bit with the version of E3SM
master that this code (and associated E3SM commit) was branched from.

When using the implicit stress calculations, rather than calculate
fluxes that are consistent with the winds provided by the atmosphere
model, the surface components calculate fluxes that are consistent with
an estimated average of the winds over the next time step, assuming that
the calculated surface stresses are applied to the wind over that
period. This is done to address numerical instability in the surface
winds, which causes large oscillations in wind speed, particularly when
the atmosphere uses a fine vertical grid near the surface, and
particularly where wind speeds are large over a particularly rough grid
cell.

In most cases, the stresses from the surface level will be rapidly
diffused away from the lowest layer of the atmosphere by the
atmosphere's boundary layer and turbulent diffusion schemes, but this
depends strongly on atmosphere's vertical grid and current state. The
implicit flux calculations therefore require the atmosphere to pass
enough information for calculations in the surface models to estimate
the response of lowest-level wind speed to the surface stress.

This update accomplishes this by passing a quantity called "wsresp" from
the atmosphere to the land. "wsresp" is the derivative of the lowest
level wind speed at the current time step with respect to the surface
stress over that time step, accounting only for the effect of vertical
diffusion. In this update, "wsresp" is only calculated to first order
with respect to the surface/atmosphere coupling time step.

This is accomplished by tracking a separate, perturbed copy of the winds
and stresses at every model level in CLUBB. These winds and stresses
evolve in the model in the same way as the "true" winds and stresses,
except that they are not allowed to affect any of other prognostic
variables, and we add a perturbation to the surface stress applied by
CLUBB to these variables. (The perturbation is currently hard-coded to
0.1 Pa). We then use the difference between the perturbed and true
lowest-level winds to estimate the derivative respresented by wsresp.

In addition to wsresp, the atmosphere must also pass an estimated
"equilibrium" value of the surface stress, i.e. a value that would
roughly balance the current forcing from the vertical diffusion and
dynamics and produce little net change in the surface winds. This value
is labelled "tau_est", and is currently estimated by simply taking the
surface stress that was applied in the previous time step. Changes to
the stress are then calculated assuming that the *difference* between
the new stress and this equilibrium value is what affects the
lowest-level wind.

As part of this commit, some shared stress calculation code for the
land, sea ice, and ocean components has been moved to shr_flux_mod. In
general, calculating the implicit surface stress requires more
iterations in loops in the various components, and conditions have been
added to test for convergence in these loops.