SCREAM: add cmake var to enable experimental code/tests
This var is off by default, so that master is always working.
Having a cmake var (and corresponding CPP macro) allows developers
to work on experimental features in an incremental way, issuing
PRs with code that is not fully working/tested instead of waiting
until it's completely working and issuing a huge PR with tons
of conflicts.