Should not set CMAKE_SYSTEM_NAME to Catamount for Intel MPI
CMake option CMAKE_SYSTEM_NAME=Catamount should only be used for
Cray compiler wrappers.

When using Intel MPI, we have to explicitly use the native Intel
compilers instead of Cray wrappers. In this case, we should not
set CMAKE_SYSTEM_NAME to Catamount.

This fix is required for E3SM to build SCORPIO with Intel MPI
on Cori.