SCREAM: revisited logic for baseline dir/ref in test-all-scream
It is now allowed to not pass a baseline dir/ref at all.
In this case, the user must be either asking for an integration test
(for which we use ref=origin/master and dir=AUTO) or a dirty tree
test (for which we use ref=HEAD and dir=ctest-build/baselines).