SCREAM: major rework of FieldManager relative to groups
Now, FM can accommodate (if possible) multiple requests for groups
to be bundled. The way it detects if it is feasible and then actuall
does it (if it is) is a bit involved. The key ingredient is the
'contiguous_superset' function in scream_utils.hpp. Using that,
the FM can detect if a bunch of mutually overlapping field groups
can be arranged so that they're all contiguous subviews of their
union.