Fix units for all physics processes so that they match.
This commit updates the units for all of the field managed variables
in the src/physics directory so that they match.  This is needed for
1. Accuracy, some units were not correct which can lead to confusion.
2. Coupling between processes, the AD will throw an error if a field
   is registered with the field manager with more than one set of units.

There is also some cosmetic changes, like fixing indentation and changing the
name of some variables to match the names we have agreed upon in our
field managed variables group discussion.