First step towards coupling all physics together.
This commit creates a new stand alone test, shoc_cld_p3_rrtmgp, which
couples SHOC, CLD-Fraction, P3 and RRTMGP together in a single run.

This commit also provides a complete input file to initialize the run
with data that was gathered from a full SCREAMv0 F90 run.

Finally, there are a couple bugfixes in the radiation interface code.

NOTE: There are also changes to the shoc interface that should not be brought
into master.  These are a quick fix to move past an initial issue where
p_int, p_mid and pseudo_density would somehow throw an error at initialization.
This bug is under investigation, but in the interest of also exploring full
physics coupling we keep this simple run around in the code, for now.