Response to reviewers 4 - fix the gas molecular weight lookup
The old version of getting the gas molecular weight was not device
friendly.  This new version should remedy that issue.

Now, instead of passing a string argument when converting between
mmr and vmr the user must pass the actual molecular weight as a Real
input.

A helper function, that only works on host, can be used to lookup the
molecular weight of a gas.  The helper function is stored in the
physics constants structure.

Tested on CPU and GPU's at LC and builds and runs.