Introduces a calculate_density function to the common physics functions framework.
The density calculation takes advantage of the definition of pseudo_density
passed to physics by the dynamics.  Psuedo_density assumes a shallow atmosphere,
i.e. cell area doesn't change with cell height.  As a result the expression for
density (rho) is a rather simple expression in terms of pseudo_density and dz.

This commit also introduces a set of new unit tests to test the calculate_density
function.

[BFB]