Allow water thickness diffusion across the GL/margin
Before this commit, the water thickness diffusion term was zeroed
at the grounding line or margin.  However, the Humboldy domain
had a lot of large water thicknesses (lakes) along the grounding line.
This commit allows water thickness diffusion to occur at the GL.
It assumes that the water thickness across the GL is zero, which can
lead to large diffusion values, but the diffusivity itself is based on
a 1-sided evaluaton of water thickness.