Ensure hydropotential adjacent to ocean does not fall below sea level
Because of the check that prevents ocean inflow in from the ocean to the glacier,
it is possible for hydropotential to drop far below the ocean level, for example if
efficient channels develop at such a location.  In effect, there is no longer an enforced
boundary condition once the hydropotential first drops below sea level, because such a
condition does not feel the b.c. via an influx from the ocean.  Thus we need to apply
a Dirichlet b.c. in those situations to keep hydrpotential close (at) sea level in
such locations.  If we ever remove the restriction that water cannot flow in from the
ocean, then this check should be reverted.