Remove LANDM_COSLAT and replace uses with LANDFRAC
Remove the LANDM variable from the COMSRF module and replace all uses
with cam_in%landfrac. LANDM was previously read in from the LANDM_COSLAT
variable in the topography dataset. This was a smoothed version of the
input topography, but was actually inconsistent with the land fraction
data specified by the surface models. Instead of carrying around two
versions of land fraction, we opt to remove the LANDM variable and
instead use a land fraction consistent with the surface models. This
will be BFB for all currently supported configurations *except* for MMF,
where we were still using the old cloud effective radius calculations
with LANDM instead of LANDFRAC. Since these are switched over to use
LANDFRAC now, the results will be non-BFB but should still be
reasonable. Also, any cases using RK microphysics will DIFF because
`stratiform_tend` was also using LANDM, buried down in the findmcnew
routine, and also in the effective radii computation. The code was
changed here but *not tested* for this configuration because we do not
have a supported configuration for that microphysics package at the
moment. Fixes 4276.