Introduce a new coupled test for homme dynamics + the full physics suite.
NOTE - BROKEN
This commit doesn't work.  The new test throws an error about qv not being
the first tracer.  This needs to be fixed.

This commit is meant to be a placeholder for debugging this issue.  So that
any changes can be reverted to this baseline during the investigation.

This commit intoduces the infrastructure for a dynamics+physics test.  It is
built off of the homme_stand_alone test and in this case only the physics
process P3 has been added.  As bugs are sorted out we will build upon this
test and add in SHOC, CLD and RRTMGP for a complete dynamics+physics test.