Remove unused nml section prescribed_ozone and use consistent oxid file
Prescribed_ozone_nl section is not needed for E3SMv2 production compsets
but still appear in some of them. There are two versions of prescribed
oxidant files for aerosol chemistry among the V2 comppsets, but they are
converted from the same netcdf4 source file. Use_case files are updated
to consistently use the one that records the detailed history of the
conversion.

[BFB] No impact on simulation results. NML diffs expected for 1850, 1950
      and 2010 WCYCL and F compsets