Include temporal interpolation - this is step 2/3
This commit introduces temporal interpolation to the SPA interface.

Currently we are still only dealing with the first 2 months of data, so
the temporal interpolation focuses on setting the CCN3 values for a time
in between months 1 and 2.

This commit interpolates the surface pressure and the aerosol concentrations
over time.

It also updates the vertical interpolation calls to use interpolated data
rather than data from the beginning of the month.