SCREAM: further limiting ability to extract 1d array from a Field
We now allow to get a pointer (rather than 1d view) out of a Field.
This will hopefully limit its usage to a handful of cases within
the infrastructure, where we can guarantee that conditions are met.

Also: removed field_monotonicity_check, since it's not clear what
the check should be. It makes no sense to check that the entries
stored (arbitrarily) in the 1d array are monotonic.