Hot-fix to initial condition files for the stand-alone tests.
The initial condition files that changed in PR #1157 were mistakenly
written for the wrong resolution which caused the homme+physics coupled
test to fail.

This commit also fixes the CMakeLists.txt file for homme+physics so that
it will be included in the AT.  Previously it was skipped, which was how
the issue was able to make it into the master code.