Homme/SL: Intel optimization_level pragma for BFBness.
On Chrysalis, despite the e3sm_atm_developer and e3sm_atm_integration passing
against baselines, I found that the PE layout test for the WC v2 problem
failed. I tracked this to a specific routine. If optimization is set to O1 for
this one routine, BFB invariance is restored, and the output is BFB with master.