Move conversion of nc_prescribed out of p3_main code and into the interface.
Previously, the nc_prescribed values passed to p3_main, gathered from SPA,
had units of #/cm3.  These needed to be converted to the units of nc, #/kg.

This was done inside of the p3_main routine, which could cause confusion in
the future if we adopt a different aerosol scheme to SPA.  Instead, the
unit conversion is now done in the p3 interface and explicitly done when
SPA is active.

This commit also introduces non-zero values for nccn_prescribed in the
p3_run_and_cmp tests.  In testing the changes in this test it was discovered
there wasn't any nccn_prescribed coverage for the p3_run_and_cmp tests.