SCREAM: fix bug in how we compute dofs offsets in scorpio input/output
The scan sum of the number of local columns should be exclusive,
but MPI_Scan performs an inclusive prefix sum. MPI also offers
MPI_Exsan, but an exclusive sum is just an inclusive sum, minus
the value on rank, so we can do this without adding the exclusive
prefix sum in ekat's Comm class.