Add damage to advection scheme
This commit adds damage to the advection scheme.  It only occurs if
config_damage_advection is .true., and it is .false. by default.  We may
want to change the namelist controls in the future, perhaps having
damage always advected when the damage calving scheme is on.  But this
makes it simpler for testing.

The advection scheme is 3d but damage is a 2d field.  When damage is
added to the tracer array, the 2d value is assigned at all vertical
levels.  After advection, the tracer array result is converted back to a
2d field by a layerThickness-weighted average.

Any accumulation at the ice surface or bottom is accumulated with a
damage value of 0.