SCREAM: bug-fix in AbstractGrid::is_unique()
The method was using memcpy, but it was using it wrong (passing the
number of items to copy, rather than the number of bytes).

This was causing not all gids to be checked, and, depending on
the gids distribution, it could cause a corrupted copy, which
at times triggered failures.

Also, added a handy getter to the grid's comm, which can be
useful while debugging.