Compute surface lw flux for standalone test
Compute surface LW flux up in standalone driver so that T_int can be
computed in run_impl, since we no longer pass T_int through the FM. Also
allow for atmosphere vertical indexing to be from surface to TOA (as the
RRTMGP standalone test uses) or TOA to surface (as SCREAM uses). This
allows us to use the same test data for the rrtmgp_standalone test as
the rrtmgp_test in src/rrtmgp/tests (which is also the same test used by
the external RRTMGP repo). Also add a missing team_barrier() after
computation of T_int. Also (last also), loosen tolerance on computed
fluxes. This was needed since computation of T_int from T_mid and
surface LW flux is not BFB with T_lev from the inputdata.