Simple fix to all AD based SHOC tests that zero out the surface fluxes.
Having a constant surface flux caused the temperature in some columns
to decrease over time at the surface because the fluxes removed energy
from the column.  This led to any test with shoc in it eventually going to
unrealistic temperature values over time.

This is a minor issue as it's impact is only felt after a significant number
of steps.  But it is an issue and exposes a error in our test setup which this
commit solves.