SCREAM: removed notion of 'updated' group in AtmProc
In order to keep FieldGroup's more parallel to Field's logic,
we make the former use required/computed semantic. It is still
possible to call 'add_group<Updated>(..)', just like it is possible
with fields, to reduce code duplication. However, the AP class does
not expose 'updated' groups, but required and computed one.