Changes made to get the new code to work on GPUs
The changes required were to map many of the device view data onto
a mirrored host view for many of the for loops.

It is possible that a low hanging fruit for performance would be to
fix many of these loops to work on device and to run in parallel.

It is worth noting though, that the remap weights are only called once
per simulation at initialization, so the performance gain may not be that
dramatic.