Improve Kokkos initialization and finalization.
When multiple components use Kokkos, we need to be careful about initializing
and finalizing Kokkos. Each can occur only once, and finalization must occur
after the last component using Kokkos has deallocated its views.

This commit modifies cime_comp_mod to finalize components in the opposite order
that it initializes them. Then a component can use is_initialized to determine
if it is in charge of initializing and finalizing Kokkos. In this commit, I've
modified SL transport to follow this pattern.