SCREAM: refactored IO class stack
The OutputManager (OM) class is now in charge of opening/closing nc files.
Each AtmosphereOutput class operates on a single grid, and does not
handle file operations. Also, the OM is in charge of updating the time
stamp in the nc file (so it gets updated once, regardless of how many
output streams write to it).
Since each nc file is handled by an OM, some code in the scorpio F90
interface was no longer needed, and was removed (e.g., calling redef
when adding new vars/dims to an nc file where enddef was already
called).