Homme/SL: Fix all remaining Cuda/GCC -Wall -pedantic warnings.
This commit removes all remaining warnings in builds with Cuda 10.1.105 and GCC
7.2.0 and 8.5.0. The remaining warnings were (1) 'warning: dynamic
initialization in unreachable code' in various spots in compose/ and (2)
'warning: calling a __host__ function from a __host__ __device__ function is not
allowed' in HommeFormatArray functions. These are benign, but the goal is to
make composec++ and composef90 build free of warnings. (This was already true in
non-Cuda builds.)