Make all tests that run through the AD run on multi-ranks.
Additionally, make it so that each run produces output labeled
by the number of mpi-ranks used.

This commit introduces a number of new tests that determine whether
or not an atmospheric process will be BFB when run with 1 or more MPI
ranks.

It introduces a new rrtmgp stand alone test that runs a set number
of steps for radiation only.  Note, this commit changes the old
rrtmgp_standalone test to the rrtmgp_standalone_unit test.

Over the course of preparing this commit a bug was found in the rrtmgp
interface which is also fixed here.

This commit establishes a CprncTest macro which will compare two netCDF
output files.

For clarity, the word stand_alone was removed from the shoc_cld_p3_rrtmgp
and homme_shoc_cld_p3_rrtmgp tests, since they are not really stand alone
tests.

Finally there is a minor change to the cmake file for lassen, and a change
to the scorpio output interface that allows a user to not have the timestamp
attached to an output file name.