Update harvest control to allow FATES to use ELM hharvest input
Configuration was updated to allow ELM harvest to read in data and pass
to FATES  when FATES is active. A PFT number check is now allowed when
FATES is active and the crop model is off in order to avoid a memory
access error. The harvest data array for reading needed an updated shape
parameter in order to be allocated correctly.
These updates do not change answers in relation to the previously
updated ELM-FATES harvest code.