SCREAM: allow adding pre-built Field's to a FieldManager
There are a few conditions, like that the FM must already be
in "closed" state, with no other field with the same name
already registered.

This capability allows to build Field's locally (so that they
can be built without the somewhat heavy FieldRequest syntax),
but still take advantage of classes that rely on FM (like
IO classes). For instance, one can create a field as a subfield
of another one, set it in the FM, and have IO output it.