SCREAM: changed bookkeping for scorpio IO files
As soon as a file is closed (no more customers of that file),
we delete it from the list of files. This makes it much more
simple to handle tests where files get reused during the same
execution. As a consequence, there's no point in checking if
a pio_atm_file that was found is open or not: if found, it's open.

Also, fixed a check when finding a variable in a file.