Makefile changes to compile with Sea Level Model
This commit modifies the Makefile to support compiling with or without
the sea level model.  To include it, add "SLM=true" on the make call.
If that is included, the build system expects the SLM submodule to have been
initialized/updated already.  It also activates a cpp directive that
includes calls to the sea level code.