Downgrading NetCDF lib from 4.8.1 to 4.8.0 for ANL GCE nodes
It has been confirmed that some E3SM cases run on ANL GCE nodes with
netcdf4p type and latest NetCDF 4.8.1 might return NC_EHDFERR from
nc_enddef(). An issue has been created for NetCDF developers to fix.

Before this issue is fixed in newer releases of NetCDF, we have to
switch back to NetCDF 4.8.0, an older but more stable release.

For consistency on GCE nodes, we downgrade the serial version of the
NetCDF lib to 4.8.0 as well.