SCREAM: allow property checks to return warnings rather than failures
It might be desirable to simply print warnings rather than crash
the simulation, in some cases. For instance, we might want to crash
the simulation if a certain field goes above 100, but would like to
know when it started to go above 90. We could then add two upper
bound property checks, one with UB=100 and one with UB=90.
The former would return Fail if the check fails, while the latter
would return Warn, which triggers a warning, but the simulation
can continue.