Only modify the mostRecentAccessTime is a stream was actually read
Before this commit, a default value of '0000-01-01_00:00:00' was being
saved to the new mostRecentAccessTime attribute on calls to mpas_stream_mgr_read()
where a stream wasn't actually read because its alarm wasn't ringing.

This commit initializes actualWhen_local to a non-time value in order to
only set mostRecentAccessTime if actualWhen_local is successfully
updated to a time by read_stream().