Bugfix to when we reset the number of nsteps since the last write.
This commit fixes a bug where we only reset the number of steps since
the last write in the output manager when an output file was full.  Since
we use this value to update any averaging output fields we want to reset
it every time we write output, even in cases where an output file can hold
more output timesnaps.