Quick fix to the default settings for scream_output.yaml
This commit fixes an issue where long CIME runs would not produce
multiple output files when a file was full.  By not storing the
timestamp in the filename, each "new" file would have the same name
as the last file and then overwrite it.

This should not be turned on by default.