Update jenkins script to better handle out-of-date PRs
If CIME tests are requested by the AT, do an upstream merge
before the CIME tests are run so that we know any DIFF is
caused by this PR and not just that the PR is old.

Use a new simple CLI hook for git_merge_ref so we can reuse
this nice functionality without having to duplicate it in bash.