SCREAM: HACK ALERT, prevent Homme CPP define from messing up YAKL
In Homme, the VECTOR_SIZE macro is defined. In YAKL's cub submodule,
an enum value named VECTOR_SIZE is declared locally in a header function.
We rearrange the headers to get the YAKL one to be (indirectly)
included first, so that it can be compiled *before* the homme macro
is expanded.