For builds using GNU v10 and higher, add a flag for one file (inidat.F90) in non-DEBUG. When updating to GNU v10 on Perlmutter, the SMS_P4x1.ne4pg2_ne4pg2.F-MMFXX test was hitting INF values in optimized builds. Tracked down that I could add -fno-inline-arg-packing to avoid -- not yet sure if it points to compiler bug.
"At any optimization level except-Os, gfortran now uses inline packing
for arguments instead of calling a library routine. If the source
contains a large number of arguments that need to be repacked, code
size or time for compilation can become excessive. If that is the
case, -fno-inline-arg-packing can be used to disable inline argument
packing. "

bfb