SCREAM: changed cmake logic for generating allsky baselines
Before, rrtmgp was generating baselines during 'make baselines',
while normal testing would use the baselines found in the baselines
folder (generated in a previous run). However, the "baselines",
are simply the output generated by raw rrtmgp calls, to be used
to compare scream::rrtmgp output against it.

So instead, we remove the custom target that generates allsky baselines
from the deps of 'make baseline', and we run it during the test
phase (if the file is not present already).