Introduce a register_diagnostics routine to register all diagnostics.
This commit introduces the register_diagnostics routine which will
use the atmosphere process factory build to register all possible
diagnostics.

This routine is then used in the scorpio_output class to be able to
create diagnostics on request.

This commit also changes the scorpio_output.cpp infrastructure to try
to retrieve a field from a diangostic if it isn't available in the
field manager.  At initialization the scorpio_output class will create
all diagnostics that it knows it will need.

The new setup is tested by making the potential_temperature diagnostic
test use `register_diagnostics` and by add a diagnostic output to the io_test.

Note, the changes to io_test are not complete.  At this commit it will run
to completion.  The test still needs to try to do soemthing meaningful and
test that the output is accurate.