Set default PIO output type to be 64bit
The default output prior to this commit was "classic".  Additionally,
we determined that the PIO output set by CIME was not being passed
to the EAMxx-scorpio interface.  This commit sets the infrastructure
to do that.

Note, there is still a bug which will need to be fixed in a subsequent
commit.

ifdef CIME_BUILD

doesn't do anything because CIME_BUILD is not a EAMxx cmake var.  Instead
it is SCREAM_CIME_BUILD.  Unfortunately fixing this exposed a separate issue
where the CIME sharedlibs needs to be added to SCREAM modules.  Since that
will be a bit more work we leave it to a subsequent commit and for now
stick with the default output type.